backend/requirements.txt
Django==4.2.9
django-debug-toolbar==4.3.0
django-environ==0.11.2
djangorestframework==3.15.0
django-filter==24.1
drf-yasg==1.21.7
psycopg2-binary==2.9.9
django-unfold==0.30.0
django-rosetta==0.10.0
django-cors-headers==4.4.0